requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.9.0
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
